+++ CHANGELOG.md
@@ -31,6 +31,9 @@ Line wrap the file at 100 chars.                                              Th
 ### Changed
 - Update the Post-Quantum secure key exchange gRPC client to use the stabilized
   `PskExchangeV1` endpoint
+- Add "auto" setting for the quantum-resistant tunnel feature, and make it the default. If it was
+  previously set to off, it will now be set to auto instead. That currently means the same thing as
+  "off", but this might change in the future.
 
 #### Windows
 - Remove automatic fallback to wireguard-go. This is done as a first step before fully